Make PSR compatible class a separate thing, adjust port handling
